refactor(outbox): promote test harness to shared file; add InMemory quirk breadcrumb
Three code-review fixes on Task 5 (7dfbd1e):

- Promote OutboxDispatcherTests.NewHarness() + WaitUntil() + JsonSerializer
  + TypeRegistry to a shared TestHarness class. Removes cross-class
  coupling and a byte-identical WaitUntil duplicate between the
  dispatcher tests and the cross-aggregate test.
- Document the magic "CONFLICT" string in IdempotencyDemoTests as a
  v0.2 deferral (StoreError.ConflictCode const). The cleanest fix
  touches a sibling package; documenting in-place keeps Task 5
  focused.
- Add a TODO breadcrumb at the InMemoryEventStoreAdapter "*"
  pseudo-stream impl pointing at the CrossAggregateIntegrationTests
  workaround. A v0.2 maintainer fixing the adapter now has a
  starting reference.

Public API surface unchanged.
